First demo commit.
This project is a PC client which send cmd and receive result from Phone side. For now, it only has a Main Form and a Command Test Form.

How to use it?
1) First compile the project
2) Choose menu 文件->开启端口转发
3) Choose menu 工具->命令测试器
4）Input test cmd, and click button "测试"

If every thing is ok, especially Phone side service is working, you will see the result returned from phone in the Commmand Test Form.